# api/openapi.yaml
openapi: 3.1.0
info:
  title: Snooze API
  version: 2.0.0
  summary: Log aggregation and alerting REST API.
  description: |
    The Snooze v1 HTTP API. Snooze 2.0 is a Go rewrite of the Python 1.x
    server; this document describes the **v1** wire contract that ships
    with 2.0.

    ## Authentication

    Every endpoint outside the small "public" set (`/healthz`,
    `/readyz`, `/metrics`, `/api/v1/login/*`) requires a bearer JWT:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `POST /api/v1/login/{local,ldap,anonymous}`.
    The Python 1.x `Authorization: JWT <token>` scheme is no longer
    accepted.

    ## Multi-tenancy

    Snooze supports multiple isolated organizations (tenants) on a single
    server. Every authenticated request is scoped to the tenant encoded in
    the `tenant_id` JWT claim. Pass the `org` field in the login request
    body to obtain a token scoped to a specific tenant; omit `org` (or leave
    it blank) to land in the `default` tenant.

    **Ingest routing** — unauthenticated alert ingestion (`POST /api/v1/alerts`,
    `POST /api/v1/webhook/*`) is routed to a tenant by the `Authorization:
    Bearer <ingest_token>` header, where `<ingest_token>` is the per-tenant
    opaque token stored in the tenant document. When no token is supplied or
    the token is unknown, the request falls back to the `default` tenant.

    **Platform permissions** — the `ro_tenant` and `rw_tenant` permissions gate
...

    ## CRUD shape
...
servers:
...
security:
  - bearerAuth: []
  # Trusted-header (auth-proxy) mode is an alternative to a Bearer token on
  # every authenticated route, but only when auth_proxy.enabled=true and the
  # immediate TCP peer is a trusted proxy. Disabled by default.
  - proxyHeader: []
...
paths:
...
  /api/v1/login/tenant:
    get:
      tags: [auth]
      summary: Resolve an opaque per-tenant login key
      description: |
        Resolves a per-tenant `login_key` to the tenant's `id` and
        `display_name`. Returns a generic 404 for an empty, unknown, or
        suspended tenant — it never resolves by slug, so it cannot be
        used to enumerate tenants. Used by the login page when the URL
        carries `?key=<login_key>`.
      security: []
      parameters:
        - name: key
          in: query
          required: true
          schema: {type: string}
          description: "The opaque per-tenant login key."
      responses:
        '200':
          description: Tenant identity for the given key.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [id, display_name]
                    properties:
                      id:
                        type: string
                        description: "Tenant slug."
                      display_name:
                        type: string
                        description: "Human-readable label."
        '404':
          description: Key not found, empty, or tenant suspended.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
...
  /api/v1/action/test:
    post:
      tags: [crud]
      summary: Send a test notification for an action config
      description: |
        Delivers a single synthetic alert through the notifier named by
        `selected`, using `subcontent` as its action_form configuration. The
        config does NOT need to be persisted first — this powers the
        "Send test" button in the Actions editor. The notifier's real Send
        path runs, so a 200 proves end-to-end delivery (URL, auth, payload).
        No record is mutated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selected]
              properties:
                selected:
                  type: string
                  description: Notifier plugin registry key (teams, slack, webhook, …).
                subcontent:
                  type: object
                  additionalProperties: true
                  description: The action_form values for that notifier.
      responses:
        '200':
          description: |
            Test notification delivered. The test send bypasses batching and
            reports the real transport outcome, so a 200 always means the
            notifier accepted the alert — there is no deferred "queued" state.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
        '404':
          description: Unknown plugin.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
        '422':
          description: Plugin is not a notifier, or upstream delivery failed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrEnvelope'}
...
  /api/v1/tenant:
    get:
      tags: [platform]
      summary: List tenants
      description: |
        Returns all tenant records. Requires `ro_tenant` or `rw_tenant`.
        Evaluated against platform scope, not any individual tenant.
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 20}
        - name: orderby
          in: query
          schema: {type: string}
        - name: asc
          in: query
          schema: {type: boolean, default: true}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  meta:
                    $ref: '#/components/schemas/ListMeta'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [platform]
      summary: Create tenant
      description: |
        Creates a new organization and seeds its default roles (`admin`,
        `viewer`, `notifications`). Requires `rw_tenant`. The `id` slug must be
        unique and must not be `default`. If `ingest_token` is omitted a random
        token is generated and returned in the response — this is the only
        opportunity to retrieve it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantCreateResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/tenant/{id}/admin:
    post:
      tags: [platform]
      summary: Ensure/reset a tenant's local admin password
      description: |
        Provisions or resets the local admin for the given tenant. If the
        user identified by `username` (default "admin") already exists its
        password is rotated; otherwise a new local user is created with the
        `admin` role. Requires `rw_tenant`. The returned `password` is shown
        exactly once — store it immediately.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
          description: "Tenant slug."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: "Admin username to provision or reset (default \"admin\")."
      responses:
        '200':
          description: Admin credential (password shown once).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCredential'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tenant/{id}/rotate-login-key:
    post:
      tags: [platform]
      summary: Rotate a tenant's login key
      description: |
        Generates a fresh opaque `login_key` for the tenant, invalidating
        any previously shared login links. Also bootstraps a key for tenants
        that were created before this feature was introduced. Requires
        `rw_tenant`. The new key is returned in the response — update any
        shared login links immediately.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
          description: "Tenant slug."
      responses:
        '200':
          description: New login key.
          content:
            application/json:
              schema:
                type: object
                required: [id, login_key]
                properties:
                  id:
                    type: string
                    description: "Tenant slug."
                  login_key:
                    type: string
                    description: "Freshly generated opaque login key."
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/tenant/{id}:
    get:
      tags: [platform]
      summary: Get tenant by id
      description: Requires `ro_tenant` or `rw_tenant`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [platform]
      summary: Update tenant display_name / status / ingest_token / listed
      description: Requires `rw_tenant`. `id` is immutable.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [platform]
      summary: Delete tenant
      description: |
        Requires `rw_tenant`. The reserved "default" tenant cannot be deleted
        (server returns 409).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        '204':
          description: Deleted
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. The Python 1.x `Authorization: JWT <token>` scheme
        is rejected. Obtain a token via `POST /api/v1/login/local`.

    ingestToken:
      type: http
      scheme: bearer
      description: |
        Per-tenant ingest token. Opaque random string stored in the tenant
        document (`ingest_token` field). Supply as `Authorization: Bearer
        <ingest_token>` or `?token=<ingest_token>`. When absent or unknown,
        the middleware falls back to the `default` tenant (D4). This scheme
        gates `/api/v1/alerts` and `/api/v1/webhook/*` in multi-tenant
        deployments; single-tenant deployments leave ingest open (existing
        behaviour preserved).

    apiKeyAuth:
      type: http
      scheme: bearer
      description: >
        A user API key, presented as `Authorization: Bearer snz_…`. Keys are
        minted at `/api/v1/user/me/apikeys`, carry a subset of the owner's
        permissions, and are bounded live by the owner's current permissions on
        every request.

    proxyHeader:
      type: apiKey
      in: header
      name: X-Forwarded-User
      description: >
        Trusted-header (auth-proxy) mode. Gated by `auth_proxy.enabled` and
        **disabled by default**. When enabled, an upstream reverse proxy
        (oauth2-proxy / Pomerium / mod_auth_openidc) that has already
        authenticated the user forwards the identity in the configured username
        header (default `X-Forwarded-User`) and an optional groups header
        (default `X-Forwarded-Groups`). Snooze JIT-provisions a `method=proxy`
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
    Forbidden:
      description: Authenticated but lacking the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
...
  schemas:
...
    Tenant:
      type: object
      required: [id, display_name, status]
      description: |
        One organization (tenant). id is the immutable URL/login-safe slug
        stamped as tenant_id on every scoped document. display_name is mutable.
      properties:
        id:
          type: string
          description: "Immutable slug, used as tenant_id on all scoped documents."
        display_name:
          type: string
          description: "Human-readable label; mutable."
        status:
          type: string
          enum: [active, suspended]
          description: "'suspended' blocks login + ingest but retains data."
        ingest_token:
          type: string
          description: "Per-tenant ingest token for unauthenticated ingress (D4)."
        listed:
          type: boolean
          default: true
          description: |
            Controls whether this tenant appears in the public login list
            returned by GET /api/v1/login. Default true. Set to false for
            SaaS tenants reached only via their per-tenant login_key link.
        login_key:
          type: string
          description: |
            Opaque, rotatable per-tenant discovery secret. Share
            /web/login?key=<login_key> to let users reach this tenant's login
            page without exposing the tenant list. Never included in the public
            login index. Rotate via POST /api/v1/tenant/{id}/rotate-login-key.
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
...
    ListMeta:
      type: object
      required: [count, limit, offset, total]
      properties:
        count:  {type: integer}
        limit:  {type: integer}
        offset: {type: integer}
        total:  {type: integer}
...
    ErrEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrBody'
...